Update 1.nugget_is_happy.rtf
Do you remember Pepita? Well, believe it or not, it also changes moods. In our Nugget model, we are going to simply represent two possible states: when she is weak and when she is happy.

And when does that happen?

Nugget is weak if her energy is less than 100.
Pepita is happy if her energy is greater than 1000.
Completing weak methods? and happy? of Pepita.

Since in this lesson we are not going to interact with the cities, we have removed everything related to them from Nugget. We just do this to make it easier for you to write the code, don't try this at home.