Create 2.alternate_reunion.rb
In Ruby, as in many other languages, this is written with the if keyword. For example:

module Joseph
  def self.arrange_room!
    self.order!
    if self.has_dirty_sheets?
      self.change_sheets!
    end
    self.make_the_bed!
  end
end
Knowing how the conditional alternative is written in Ruby, we want Pepita, in addition to receiving orders, to have her moments to be able to do what she wants.

Obviously, what you want to do at any given moment depends on your mood:

If he is weak, he eats ten grams of canary seed to recover.
If it isn't, it doesn't do anything.
Make Pepita understand the message do_what_you_want! that behave as we explained.