Create 3.Let's_repeat_what_if_not.rb
module Gardener
  def self.care!(plant)
    if plant.needs_water?
      3.times { self.water! plant }
    else
      self.remove_bugs! plant
    end
  end
end
And what is that times?

It is a message that numbers understand that is used to execute a piece of code several times. In this case we would water the plant received as an argument 3 times.

Now that we know about the existence of times and we have seen how to do else...

Modify the solution so that if Pepita is not weak, she flies in circles 3 times.