Create 5.Pepo_arrived.rb
Pepo is a sparrow who also knows how to eat, fly and do whatever he wants, but he does it differently from Pepita.

eat birdseed: Pepo's digestive system isn't doing very well, that's why he can only take advantage of half of the birdseed he eats. For example, if you eat 20 grams of birdseed, your energy only increases by 10.
fly in circles: spends 15 energy units if it is heavy and 5 if it is not. We say that it is heavy if its energy is greater than 1100.
do what he wants: as he is always hungry, he takes advantage and eats 120 grams of birdseed.
Oh, and just like Nugget, her energy starts at 1000.

I defined Pepo according to the above rules. We leave you the Pepita code to use as a base, modify and delete the parts that do not correspond.