Create 6.a_train.rb
Our birds want to perform at the next Olympics, and for that they need to exercise a little.

To help us in this task we got Pachorra, a former soccer coach who now works with birds. He designed a special routine consisting of the following:

Fly in circles 10 times.
Eat a handful of 30 grams of birdseed.
Fly in circles 5 times.
As a reward, let the bird do what it wants.
Create Pachorra, the bird trainer, and when you receive the message train_bird! have Pepita perform her routine (yes, you can only enter Pepita , but we'll fix that soon).