Create Going_into_heat.rb
Let's create a video game library! To begin with, we will have three video games, of which we know the following:

CarlosDuty: is violent. Its difficulty is calculated as 30 - @amount_achievements * 0.5. And if it is played for more than 2 hours in a row, an achievement is added to its amount. Initially, the game has no achievements.
TimbaElLeon: It's not violent. Its initial difficulty is 25 and it grows one point for each hour played.
Metroid: Violent only if @spatial_level is greater than 5. This level starts at 3 but increases by 1 each time it is played, no matter how long. Also, its difficulty is always 100.
I defined these three objects in such a way that they understand the messages difficulty, violent? and play!(a_time).