Create 2.Creating_a_list.rb
Now that we have our video games, we are going to order them somewhere.

For this we need to define an object, the Game Library, which contains other objects: our video games. For this we are going to use a list of objects: it is a type of collection in which the elements can be repeated. That is, the same object can appear more than once.

For example, the list of numbers 2, 3, 3, and 9 is written like this:

[2, 3, 3, 9]
Let's see if this is understood: I defined a Game Library object that has a games attribute with its corresponding getter. The Game Library must have the CarlosDuty game first, then TimbaElLeon and lastly Metroide.