Create 3.Improving_the_Game_Library.rb
We took care of the video games first, and now you know what messages the lists understand. It's time to give functionality to the Game Library!

Our Game Library manages points. Add the necessary code so that it understands the following messages:

points: it tells us how many points the Game Library has. Initially they are 0.
acquire_game!(a_game): adds the game to the Game Library, and adds 150 points.
delete_game!(one_game): removes a game from the Game Library, but does not subtract points.
complete?: it is fulfilled if the Game Library has more than 1000 points and more than 5 games.
recommended_game?(a_game): is it true for a_game if it is not in the Game Library and is violent?.