Create 4.Blocks_with_parameters.rb
Blocks can also receive arguments to your application. For example, add_to_other_two takes two arguments, enclosed in pipes | and separated by commas:

a_number = 3
add_other_two = proc { |one_add, another_add| one_number = one_number + one_add + another_add }
To apply the add_to_other_two block, the desired arguments are passed to the call message:

> add_two_others.call(1,2)
=> 6
Let's go back to video games...

In the Library you can see the TimbaElLeon object. To solve this exercise, we are not interested in how the methods of this object are defined, we just want to remember the messages that it understands, that's why there are ellipses!

Assign the variable play_a_timba a block that receives a single parameter. The block receives a number of minutes and must make TimbaElLeon play during that time, but remember to play! wait a number of hours.