Create 5.Filtering_those_who_comply.rb
What happens when we want all those objects that meet a certain condition in a certain collection? For example, if from a list of numbers we want those greater than 3.

What we use is the select message from collections. select receives a block with a parameter that represents an element of the collection and a boolean condition as code, and what it returns is a new collection with the elements that satisfy it.

some_numbers = [1, 2, 3, 4, 5]
greater_than_3 = some_numbers.select { |a_number| a_number > 3 }
And when is that block that receives the select applied? The select is the one who decides! The collection will apply it with each of the objects (a_number) when appropriate during the selection (or filtering) of elements.

> older_than_3
=> [4, 5]
Meanwhile, in our game library...

Now it's your turn! I defined the violent_games method that returns the games from the Game Library that meet violent?.