Create 6.If_you_search_you_will_find.rb
What if instead of all the elements that meet a condition, we only want one? We use find!

some_numbers = [1, 2, 3, 4, 5]
one_greater_than_3 = some_numbers.find { |a_number| a_number > 3 }
While select returns a collection, find returns only one element.

> one_greater_than_3
=> 4
What if no element in the collection meets the condition? Returns nil, which, as you learned earlier, is an object that represents nothing - or in this case, that neither meets the condition.

Let's see if it's understood: make the Game Library understand game_more_difficult_than(a_difficulty), which returns some game in the Game Library with more difficulty than the one passed as an argument.