Create 7.someone_meets.rb
To find out if all the elements of a collection meet a certain criteria, we can use the all? message, which also receives a block. For example, if we have a collection of students, we can know if the whole group passes in the following way:

students.all? { |a_student| a_student.approved? }
In a very similar way, we can know if any element of the collection meets a certain condition through the message any?. Following the previous example, we now want to know if at least someone passed:

students.any? { |a_student| a_student.approved? }
I defined the following methods in our Game Library:

much_violence?: It is fulfilled if all the games it owns are violent.
very_difficult?: tells us if any of the games has more than 25 difficulty points.