Create 8.dear_map.rb
The map message allows us, from a collection, to obtain another collection with each of the results that returns a message sent to each element.

In other words, the new collection will have what is returned by the message that is sent to each of the elements. For example, if we use map to find out the energy levels of a collection of swallows:

> [Pepita, Norita].map { |a_swallow| a_swallow.energy }
=> [77, 52]
Like the rest of the messages we've seen so far, map doesn't modify the original collection or its elements, but instead returns a new collection.

Add to the Game Library a method called violent_difficulty that returns a collection with the difficulty of its violent_games.