Create 9.how_much_do_they_add.rb
Returning to our student collection. We already asked if the whole group passed or if at least someone passed using all? and any?. What if we want to know how many passed? We use count:

students.count { |a_student| a_student.approved? }
count tells us how many elements of a collection meet the condition. On the other hand, to calculate sums we have the message sum. If we want to know the sum of all the grades of the student collection, for example, we can do:

students.sum { |a_student| a_student.grade_in_exam }
Let's see if it is understood: add to the Game Library the average_de_violence method, whose value is the sum of the difficulty of the violent games divided by the number of violent games in the Game Library.