Create 10.playing_everything.rb
So far, all the messages we've seen from collections (with the exception of push and delete) are not intended to have system effects. What happens, then, when we want to do something with each element? Unlike the map, we are not interested in the results of sending the same message to each object, but rather sending a message to each one with the intention of producing an effect.

It is in this case that the each message is useful.

For example, if we want from a collection of swallows, those with energy greater than 100 to fly to Iruya, we could combine select and each to do:

swallows
  .select { |a_swallow| a_swallow.energy > 100 }
  .each { |a_swallow| a_swallow.fly_towards! Iruya }
Since we almost finished the guide and taking advantage of the fact that we have a collection of video games, what we want is... to play them all!

I defined the method play_all! in the Game Library, to play each of the games for 5 hours. Remember that games mean playing! (a_time).