Create 2.attacking_zombie.rb
We present to you the first of the survivors of the invasion, Juliana. For now his behavior is simple: he knows how to attack! a zombie with a certain number of damage points. And by doing so, the zombie takes damage.

It also has an energy level, which starts at 1000, but we won't do anything with it yet. I defined a getter method for this attribute.

Let's see if it's understood: I defined the object Juliana that can attack! on a zombie by making it receive_damage!, and initialize its energy to 1000.