Create 4.Initializing_instances.eb
As you just saw, health is not shared between bouba and kiki even though they are both Zombie instances.

But we were left with a mysterious method to clear up: initialize. When working with classes we have to initialize the attributes somewhere. That's what that method is for!

The initialize message allows us to specify how we want the instance of a class to be initialized. It's that easy!

Anastasia arrives to fight the zombies! I defined a Survivor class that knows how to attack! zombies and initialize the energy to 1000. In the Library tab you can see the code of the original Juliana.

Next, I defined juliana and anastasia as instances of the new Survivor class.